requirements-test.txt
-r requirements.txt

black==20.8b1
factory_boy==3.1.0
flake8==3.8.4
pytest-factoryboy==2.0.3
# mypy version supported my django stubs at the moment
mypy==0.782
django-stubs==1.6.0
djangorestframework-stubs==1.2.0
drf-yasg-stubs==0.1.2
pre-commit==2.7.1
pytest==6.1.1
pytest-eradicate==0.0.5
pytest-django==3.10.0
pytest-mock==3.5.1